app/requirements.txt
Flask==3.1.0
protobuf==5.28.3
python_bcrypt==0.3.2
pytz==2023.3.post1
SQLAlchemy==2.0.36
